Refactored chat and message data sources into a single class. Allows passing of the current data source to the message presenter.
