+++ src/main/java/catchat/messages/MessagesPresenter.java
@@ -1,11 +1,7 @@
 package catchat.messages;
 
 import catchat.data.entities.chat.Chat;
-import catchat.data.entities.message.Message;
 import catchat.data.source.chats.ChatDataSource;
-import catchat.data.source.messages.MessageDataSource;
-
-import java.util.List;
 
 /**
  * Created by andrew on 4/16/18.
@@ -14,8 +10,7 @@ public class MessagesPresenter implements MessagesContract.Presenter, ChatDataSo
     private ChatDataSource dataSource;
     private MessagesContract.View view;
 
-    public MessagesPresenter(ChatDataSource dataSource, MessagesContract.View view) {
-        this.dataSource = dataSource;
+    public MessagesPresenter(MessagesContract.View view) {
         this.view = view;
     }
 
@@ -41,5 +36,10 @@ public class MessagesPresenter implements MessagesContract.Presenter, ChatDataSo
     @Override
     public void onChatLoaded(Chat chat) {
         view.showChatDetails(chat);
+    }
+
+    @Override
+    public void setDataSource(ChatDataSource dataSource) {
+        this.dataSource = dataSource;
     }
 }
